Revert "Feature/rename logging db"
