+++ internal/infrastructure/rdb/driver/loggingdb/README.md
@@ -1,19 +1,19 @@
-# logging driver
+# loggingdb
 
 English | [日本語](README.ja.md)
 
 Overview: **An Observability wrapper layer that adds SQL execution logs and trace information to DB access. Actual query execution is delegated to the driver layer, and only log formatting and trace integration are added.**
 
-logging driver is an **observability adapter positioned above the driver**.
+loggingdb is an **observability adapter positioned above the driver**.
 
 ## Architectural Position
 
 ```mermaid
 flowchart TB
-    Usecase --> Repo["Repository"] --> Logging["logging driver"] --> Driver["driver"] --> DB["PostgreSQL"]
+    Usecase --> Repo["Repository"] --> Logging["loggingdb"] --> Driver["driver"] --> DB["PostgreSQL"]
 ```
 
-logging driver **does not perform DB execution processing itself.**
+loggingdb **does not perform DB execution processing itself.**
 
 It wraps `driver.DBTX` and adds the following processing during SQL execution:
 
@@ -44,11 +44,11 @@ do not need to be aware of these at all.
 
 ## DBTX Wrapper
 
-The core implementation of logging driver is a **DBTX wrapper**.
+The core implementation of loggingdb is a **DBTX wrapper**.
 
 ```mermaid
 flowchart TB
-    DBTX["driver.DBTX"] --> Wrap["wrap"] --> Logging["logging driver (dbWithLogging)"] --> Obs["SQL logging + tracing"]
+    DBTX["driver.DBTX"] --> Wrap["wrap"] --> Logging["loggingdb (dbWithLogging)"] --> Obs["SQL logging + tracing"]
 ```
 
 `dbWithLogging` wraps `driver.DBTX` and performs the following processing before and after SQL execution:
@@ -83,7 +83,7 @@ to be **traced within a single trace context**.
 
 ## Slow Query
 
-logging driver automatically determines slow queries.
+loggingdb automatically determines slow queries.
 
 The determination of slow queries depends on the following configuration.
 
@@ -102,7 +102,7 @@ flowchart TB
 
 ## Provider
 
-`DBProvider` is a **DI Adapter** that aggregates the dependencies required for logging driver.
+`DBProvider` is a **DI Adapter** that aggregates the dependencies required for loggingdb.
 
 Provided dependencies:
 
@@ -112,7 +112,7 @@ Provided dependencies:
 - `DatabaseConfig`
 - `LayerTracer`
 
-This allows logging driver to:
+This allows loggingdb to:
 
 - logging implementation
 - tracing implementation
@@ -154,15 +154,15 @@ Therefore, it is very effective during development.
 
 ## Notes
 
-### logging driver does not perform DB I/O
+### loggingdb does not perform DB I/O
 
-logging driver is a **pure wrapper**.
+loggingdb is a **pure wrapper**.
 
 All actual SQL execution is delegated to the driver layer.
 
 ```mermaid
 flowchart TB
-    Logging["logging driver"] --> Driver["driver"]
+    Logging["loggingdb"] --> Driver["driver"]
 ```
 
 ### Always propagate Context
