+++ internal/infrastructure/rdb/driver/loggingdb/mock/mock_provider.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=provider.go -destination=mock/mock_provider.go -package=mock_loggingdriver
+//	mockgen -source=provider.go -destination=mock/mock_provider.go -package=mock_loggingdb
 //
 
-// Package mock_loggingdriver is a generated GoMock package.
-package mock_loggingdriver
+// Package mock_loggingdb is a generated GoMock package.
+package mock_loggingdb
 
 import (
 	config "boilerplate-go/internal/config"
