+++ internal/infrastructure/rdb/driver/loggingdb/provider_test.go
@@ -1,4 +1,4 @@
-package loggingdriver
+package loggingdb
 
 import (
 	"context"
