+++ internal/infrastructure/rdb/driver/loggingdb/with_logging.go
@@ -1,4 +1,4 @@
-package loggingdriver
+package loggingdb
 
 import (
 	"context"
