+++ internal/infrastructure/rdb/driver/loggingdb/with_logging_test.go
@@ -1,4 +1,4 @@
-package loggingdriver
+package loggingdb
 
 import (
 	"context"
@@ -8,7 +8,7 @@ import (
 
 	"boilerplate-go/internal/config"
 	"boilerplate-go/internal/infrastructure/rdb/driver"
-	mock_loggingdb "boilerplate-go/internal/infrastructure/rdb/driver/loggingdriver/mock"
+	mock_loggingdb "boilerplate-go/internal/infrastructure/rdb/driver/loggingdb/mock"
 	mock_driver "boilerplate-go/internal/infrastructure/rdb/driver/mock"
 	"boilerplate-go/internal/logging"
 	mock_logging "boilerplate-go/internal/logging/mock"
